Refactor ovtracer.py and update version
