+++ bin/openvisualizer_app.py
@@ -5,8 +5,7 @@
 # https://openwsn.atlassian.net/wiki/display/OW/License
 
 """
-Contains application model for OpenVisualizer. Expects to be called by
-top-level UI module.  See main() for startup use.
+Contains application model for OpenVisualizer. Expects to be called by top-level UI module.  See main() for startup use.
 """
 
 import json
@@ -18,6 +17,7 @@ import signal
 import sys
 
 import utils as u
+from openvisualizer.ovtracer import OVtracer
 from openvisualizer.SimEngine import SimEngine, MoteHandler
 from openvisualizer.eventbus import eventbusmonitor
 from openvisualizer.jrc import jrc
@@ -34,8 +34,7 @@ log = logging.getLogger('openVisualizerApp')
 
 class OpenVisualizerApp(object):
     """
-    Provides an application model for OpenVisualizer. Provides common,
-    top-level functionality for several UI clients.
+    Provides an application model for OpenVisualizer. Provides common, top-level functionality for several UI clients.
     """
 
     def __init__(self, conf_dir, data_dir, log_dir, simulator_mode, num_motes, trace, debug, use_page_zero,
