+++ bin/webserver.py
@@ -14,7 +14,7 @@ import time
 import bottle
 from bottle import view, response
 
-from openvisualizer import ovVersion
+from openvisualizer import version
 from openvisualizer.BspEmulator import VcdLogger
 from openvisualizer.SimEngine import SimEngine
 from openvisualizer.eventbus.eventbusclient import EventBusClient
@@ -23,7 +23,7 @@ from openvisualizer.motehandler.motestate.motestate import MoteState
 log = logging.getLogger('openVisualizerWeb')
 
 # add default parameters to all bottle templates
-view = functools.partial(view, ovVersion='.'.join(list([str(v) for v in ovVersion.VERSION])))
+view = functools.partial(view, ovVersion='.'.join(list([str(v) for v in version.VERSION])))
 
 
 class WebServer(EventBusClient):
