+++ openvisualizer/version.py
@@ -1,5 +1,5 @@
 MAJOR = 1
-MINOR = 5
+MINOR = 6
 PATCH = 2
 
 VERSION = (MAJOR,MINOR,PATCH)
